% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubviz_method.R
\name{estimate}
\alias{estimate}
\alias{Estimate.hubviz}
\title{Return hubviz result}
\usage{
estimate(object)
}
\arguments{
\item{object}{Output of hubviz function}
}
\description{
Return estimates and MCMC results
}
