
\name{hpdplot}
\alias{hpdplot}
\alias{hpdplot.hubviz}
\docType{methods}

\title{HPD plot}

\usage{
\S4method{hpdplot}{hubviz}(object)
}

\arguments{
\item{object}{Output of hubviz function}

\item{xlim}{X coordinate vector, (min,max)}

\item{ylim}{Y coordinate vector, (min,max)}
}
\description{
Return highest posterior density (HPD) intervals for latent positions in an MCMC and bubble plot
}

