% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{hubviz-class}
\alias{hubviz-class}
\title{An S4 class to represent hubviz model fitting results.}
\description{
An S4 class to represent hubviz model fitting results.
}
\section{Slots}{

\describe{
\item{\code{data}}{data}

\item{\code{init}}{initial value}

\item{\code{result}}{result}
}}

