% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubviz_method.R
\docType{methods}
\name{plot,hubviz,missing-method}
\alias{plot,hubviz,missing-method}
\alias{plot.hubviz}
\title{Plotting function}
\usage{
\S4method{plot}{hubviz,missing}(x, y, xlim = NA, ylim = NA)
}
\arguments{
\item{x}{Output of hubviz function}

\item{xlim}{X coordinate vector, (min,max)}

\item{ylim}{Y coordinate vector, (min,max)}
}
\description{
Hub-centric visualization
}
