% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubviz_method.R
\docType{methods}
\name{show,hubviz-method}
\alias{show,hubviz-method}
\alias{show.hubviz}
\title{Summary of hubviz results}
\usage{
\S4method{show}{hubviz}(object)
}
\arguments{
\item{object}{Output of hubviz function}
}
\description{
Summary of hubviz results
}
