
\name{plot,hubviz,missing-method}
\alias{plot,hubviz,missing-method}
\alias{plot.hubviz}
\docType{methods}

\title{Plotting function}

\usage{
\S4method{plot}{hubviz,missing}(x, y, xlim = NA, ylim = NA)
}

\arguments{
\item{x}{Output of hubviz function}

\item{xlim}{X coordinate vector, (min,max)}

\item{ylim}{Y coordinate vector, (min,max)}
}

\description{
Hub-centric visualization
}
